feat: [CDS-101927]: Categorize CD Artifact Repositories Connectors
