+++ internal/provider/provider.go
@@ -49,6 +49,7 @@ import (
 	cd_trigger "github.com/harness/terraform-provider-harness/internal/service/cd/trigger"
 	"github.com/harness/terraform-provider-harness/internal/service/cd/user"
 	"github.com/harness/terraform-provider-harness/internal/service/cd/yamlconfig"
+	cdng_connector_artifactRepositories "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/connector/artifactRepositories"
 	cdng_connector_cloudProviders "github.com/harness/terraform-provider-harness/internal/service/cd_nextgen/connector/cloudProviders"
 	pl_account "github.com/harness/terraform-provider-harness/internal/service/platform/account"
 	pl_apikey "github.com/harness/terraform-provider-harness/internal/service/platform/api_key"
@@ -164,7 +165,7 @@ func Provider(version string) func() *schema.Provider {
 				"harness_platform_connector_azure_cloud_cost":      connector.DataSourceConnectorAzureCloudCost(),
 				"harness_platform_connector_appdynamics":           connector.DatasourceConnectorAppDynamics(),
 				"harness_platform_connector_elasticsearch":         connector.DatasourceConnectorElasticSearch(),
-				"harness_platform_connector_artifactory":           connector.DatasourceConnectorArtifactory(),
+				"harness_platform_connector_artifactory":           cdng_connector_artifactRepositories.DatasourceConnectorArtifactory(),
 				"harness_platform_connector_aws_secret_manager":    pl_secretManagers.DatasourceConnectorAwsSM(),
 				"harness_platform_connector_aws":                   cdng_connector_cloudProviders.DatasourceConnectorAws(),
 				"harness_platform_connector_awscc":                 connector.DatasourceConnectorAwsCC(),
@@ -172,7 +173,7 @@ func Provider(version string) func() *schema.Provider {
 				"harness_platform_connector_bitbucket":             connector.DatasourceConnectorBitbucket(),
 				"harness_platform_connector_customhealthsource":    connector.DatasourceConnectorCustomHealthSource(),
 				"harness_platform_connector_datadog":               connector.DatasourceConnectorDatadog(),
-				"harness_platform_connector_docker":                connector.DatasourceConnectorDocker(),
+				"harness_platform_connector_docker":                cdng_connector_artifactRepositories.DatasourceConnectorDocker(),
 				"harness_platform_connector_jdbc":                  connector.DatasourceConnectorJDBC(),
 				"harness_platform_connector_dynatrace":             connector.DatasourceConnectorDynatrace(),
 				"harness_platform_connector_gcp":                   cdng_connector_cloudProviders.DatasourceConnectorGcp(),
@@ -180,12 +181,12 @@ func Provider(version string) func() *schema.Provider {
 				"harness_platform_connector_git":                   connector.DatasourceConnectorGit(),
 				"harness_platform_connector_github":                connector.DatasourceConnectorGithub(),
 				"harness_platform_connector_gitlab":                connector.DatasourceConnectorGitlab(),
-				"harness_platform_connector_helm":                  connector.DatasourceConnectorHelm(),
-				"harness_platform_connector_oci_helm":              connector.DatasourceConnectorOciHelm(),
+				"harness_platform_connector_helm":                  cdng_connector_artifactRepositories.DatasourceConnectorHelm(),
+				"harness_platform_connector_oci_helm":              cdng_connector_artifactRepositories.DatasourceConnectorOciHelm(),
 				"harness_platform_connector_jira":                  connector.DatasourceConnectorJira(),
-				"harness_platform_connector_jenkins":               connector.DataSourceConnectorJenkins(),
+				"harness_platform_connector_jenkins":               cdng_connector_artifactRepositories.DataSourceConnectorJenkins(),
 				"harness_platform_connector_kubernetes":            cdng_connector_cloudProviders.DatasourceConnectorKubernetes(),
-				"harness_platform_connector_nexus":                 connector.DatasourceConnectorNexus(),
+				"harness_platform_connector_nexus":                 cdng_connector_artifactRepositories.DatasourceConnectorNexus(),
 				"harness_platform_connector_pagerduty":             connector.DatasourceConnectorPagerDuty(),
 				"harness_platform_connector_prometheus":            connector.DatasourceConnectorPrometheus(),
 				"harness_platform_connector_rancher":               cdng_connector_cloudProviders.DatasourceConnectorRancher(),
@@ -301,7 +302,7 @@ func Provider(version string) func() *schema.Provider {
 				"harness_platform_connector_azure_cloud_cost":      connector.ResourceConnectorAzureCloudCost(),
 				"harness_platform_connector_appdynamics":           connector.ResourceConnectorAppDynamics(),
 				"harness_platform_connector_elasticsearch":         connector.ResourceConnectorElasticSearch(),
-				"harness_platform_connector_artifactory":           connector.ResourceConnectorArtifactory(),
+				"harness_platform_connector_artifactory":           cdng_connector_artifactRepositories.ResourceConnectorArtifactory(),
 				"harness_platform_connector_aws_secret_manager":    pl_secretManagers.ResourceConnectorAwsSM(),
 				"harness_platform_connector_aws":                   cdng_connector_cloudProviders.ResourceConnectorAws(),
 				"harness_platform_connector_awscc":                 connector.ResourceConnectorAwsCC(),
@@ -309,7 +310,7 @@ func Provider(version string) func() *schema.Provider {
 				"harness_platform_connector_bitbucket":             connector.ResourceConnectorBitbucket(),
 				"harness_platform_connector_customhealthsource":    connector.ResourceConnectorCustomHealthSource(),
 				"harness_platform_connector_datadog":               connector.ResourceConnectorDatadog(),
-				"harness_platform_connector_docker":                connector.ResourceConnectorDocker(),
+				"harness_platform_connector_docker":                cdng_connector_artifactRepositories.ResourceConnectorDocker(),
 				"harness_platform_connector_jdbc":                  connector.ResourceConnectorJDBC(),
 				"harness_platform_connector_dynatrace":             connector.ResourceConnectorDynatrace(),
 				"harness_platform_connector_gcp":                   cdng_connector_cloudProviders.ResourceConnectorGcp(),
@@ -317,13 +318,13 @@ func Provider(version string) func() *schema.Provider {
 				"harness_platform_connector_git":                   connector.ResourceConnectorGit(),
 				"harness_platform_connector_github":                connector.ResourceConnectorGithub(),
 				"harness_platform_connector_gitlab":                connector.ResourceConnectorGitlab(),
-				"harness_platform_connector_helm":                  connector.ResourceConnectorHelm(),
-				"harness_platform_connector_oci_helm":              connector.ResourceConnectorOciHelm(),
+				"harness_platform_connector_helm":                  cdng_connector_artifactRepositories.ResourceConnectorHelm(),
+				"harness_platform_connector_oci_helm":              cdng_connector_artifactRepositories.ResourceConnectorOciHelm(),
 				"harness_platform_connector_jira":                  connector.ResourceConnectorJira(),
-				"harness_platform_connector_jenkins":               connector.ResourceConnectorJenkins(),
+				"harness_platform_connector_jenkins":               cdng_connector_artifactRepositories.ResourceConnectorJenkins(),
 				"harness_platform_connector_kubernetes":            cdng_connector_cloudProviders.ResourceConnectorK8s(),
 				"harness_platform_connector_newrelic":              connector.ResourceConnectorNewRelic(),
-				"harness_platform_connector_nexus":                 connector.ResourceConnectorNexus(),
+				"harness_platform_connector_nexus":                 cdng_connector_artifactRepositories.ResourceConnectorNexus(),
 				"harness_platform_connector_pagerduty":             connector.ResourceConnectorPagerDuty(),
 				"harness_platform_connector_prometheus":            connector.ResourceConnectorPrometheus(),
 				"harness_platform_connector_rancher":               cdng_connector_cloudProviders.ResourceConnectorK8sRancher(),
