+++ internal/service/cd_nextgen/connector/artifactRepositories/artifactory_data_source_test.go
@@ -1,4 +1,4 @@
-package connector_test
+package artifactRepositories_test
 
 import (
 	"fmt"
