+++ internal/service/cd_nextgen/connector/artifactRepositories/jenkins_data_source.go
@@ -1,4 +1,4 @@
-package connector
+package artifactRepositories
 
 import (
 	"context"
