+++ internal/service/cd_nextgen/connector/artifactRepositories/nexus_data_source.go
@@ -1,18 +1,22 @@
-package connector
+package artifactRepositories
 
 import (
+	"fmt"
+	"strings"
+
+	"github.com/harness/harness-go-sdk/harness/nextgen"
 	"github.com/harness/terraform-provider-harness/helpers"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func DatasourceConnectorHelm() *schema.Resource {
+func DatasourceConnectorNexus() *schema.Resource {
 	resource := &schema.Resource{
-		Description: "Datasource for looking up a HTTP Helm connector.",
-		ReadContext: resourceConnectorHelmRead,
+		Description: "Datasource for looking up a Nexus connector.",
+		ReadContext: resourceConnectorNexusRead,
 
 		Schema: map[string]*schema.Schema{
 			"url": {
-				Description: "URL of the helm server.",
+				Description: "URL of the Nexus server.",
 				Type:        schema.TypeString,
 				Computed:    true,
 			},
@@ -21,6 +25,11 @@ func DatasourceConnectorHelm() *schema.Resource {
 				Type:        schema.TypeSet,
 				Computed:    true,
 				Elem:        &schema.Schema{Type: schema.TypeString},
+			},
+			"version": {
+				Description: fmt.Sprintf("Version of the Nexus server. Valid values are %s", strings.Join(nextgen.NexusVersionSlice, ", ")),
+				Type:        schema.TypeString,
+				Computed:    true,
 			},
 			"credentials": {
 				Description: "Credentials to use for authentication.",
