+++ internal/service/cd_nextgen/connector/artifactRepositories/nexus_data_source_test.go
@@ -1,37 +1,36 @@
-package connector_test
+package artifactRepositories_test
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/harness/harness-go-sdk/harness/nextgen"
 	"github.com/harness/harness-go-sdk/harness/utils"
 	"github.com/harness/terraform-provider-harness/internal/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 )
 
-func TestAccDataSourceConnectorOciHelm(t *testing.T) {
+func TestAccDataSourceConnectorNexus(t *testing.T) {
 
 	var (
 		name         = fmt.Sprintf("%s_%s", t.Name(), utils.RandStringBytes(4))
-		resourceName = "data.harness_platform_connector_oci_helm.test"
+		resourceName = "data.harness_platform_connector_nexus.test"
 	)
 
 	resource.UnitTest(t, resource.TestCase{
 		PreCheck:          func() { acctest.TestAccPreCheck(t) },
 		ProviderFactories: acctest.ProviderFactories,
-		ExternalProviders: map[string]resource.ExternalProvider{
-			"time": {},
-		},
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDataSourceConnectorOciHelm(name),
+				Config: testAccDataSourceConnectorNexus(name),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "id", name),
 					resource.TestCheckResourceAttr(resourceName, "identifier", name),
 					resource.TestCheckResourceAttr(resourceName, "name", name),
 					resource.TestCheckResourceAttr(resourceName, "description", "test"),
 					resource.TestCheckResourceAttr(resourceName, "tags.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "url", "admin.azurecr.io"),
+					resource.TestCheckResourceAttr(resourceName, "version", nextgen.NexusVersions.V3X.String()),
+					resource.TestCheckResourceAttr(resourceName, "url", "https://nexus.example.com"),
 					resource.TestCheckResourceAttr(resourceName, "delegate_selectors.#", "1"),
 				),
 			},
@@ -39,41 +38,25 @@ func TestAccDataSourceConnectorOciHelm(t *testing.T) {
 	})
 }
 
-func testAccDataSourceConnectorOciHelm(name string) string {
+func testAccDataSourceConnectorNexus(name string) string {
 	return fmt.Sprintf(`
-	resource "harness_platform_secret_text" "test" {
-		identifier = "%[1]s"
-		name = "%[1]s"
-		description = "test"
-		tags = ["foo:bar"]
-
-		secret_manager_identifier = "harnessSecretManager"
-		value_type = "Inline"
-		value = "secret"
-	}
-
-		resource "harness_platform_connector_oci_helm" "test" {
+		resource "harness_platform_connector_nexus" "test" {
 			identifier = "%[1]s"
 			name = "%[1]s"
 			description = "test"
 			tags = ["foo:bar"]
 
-			url = "admin.azurecr.io"
+			url = "https://nexus.example.com"
 			delegate_selectors = ["harness-delegate"]
+			version = "3.x"
 			credentials {
 				username = "admin"
-				password_ref = "account.${harness_platform_secret_text.test.id}"
+				password_ref = "account.doNotDeleteHSM"
 			}
-			depends_on = [time_sleep.wait_4_seconds]
 		}
 
-		resource "time_sleep" "wait_4_seconds" {
-			depends_on = [harness_platform_secret_text.test]
-			destroy_duration = "4s"
-		}
-
-		data "harness_platform_connector_oci_helm" "test" {
-			identifier = harness_platform_connector_oci_helm.test.identifier
+		data "harness_platform_connector_nexus" "test" {
+			identifier = harness_platform_connector_nexus.test.identifier
 		}
 	`, name)
 }
