+++ internal/service/cd_nextgen/connector/artifactRepositories/oci_helm_data_source_test.go
@@ -1,36 +1,37 @@
-package connector_test
+package artifactRepositories_test
 
 import (
 	"fmt"
 	"testing"
 
-	"github.com/harness/harness-go-sdk/harness/nextgen"
 	"github.com/harness/harness-go-sdk/harness/utils"
 	"github.com/harness/terraform-provider-harness/internal/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 )
 
-func TestAccDataSourceConnectorNexus(t *testing.T) {
+func TestAccDataSourceConnectorOciHelm(t *testing.T) {
 
 	var (
 		name         = fmt.Sprintf("%s_%s", t.Name(), utils.RandStringBytes(4))
-		resourceName = "data.harness_platform_connector_nexus.test"
+		resourceName = "data.harness_platform_connector_oci_helm.test"
 	)
 
 	resource.UnitTest(t, resource.TestCase{
 		PreCheck:          func() { acctest.TestAccPreCheck(t) },
 		ProviderFactories: acctest.ProviderFactories,
+		ExternalProviders: map[string]resource.ExternalProvider{
+			"time": {},
+		},
 		Steps: []resource.TestStep{
 			{
-				Config: testAccDataSourceConnectorNexus(name),
+				Config: testAccDataSourceConnectorOciHelm(name),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "id", name),
 					resource.TestCheckResourceAttr(resourceName, "identifier", name),
 					resource.TestCheckResourceAttr(resourceName, "name", name),
 					resource.TestCheckResourceAttr(resourceName, "description", "test"),
 					resource.TestCheckResourceAttr(resourceName, "tags.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "version", nextgen.NexusVersions.V3X.String()),
-					resource.TestCheckResourceAttr(resourceName, "url", "https://nexus.example.com"),
+					resource.TestCheckResourceAttr(resourceName, "url", "admin.azurecr.io"),
 					resource.TestCheckResourceAttr(resourceName, "delegate_selectors.#", "1"),
 				),
 			},
@@ -38,25 +39,41 @@ func TestAccDataSourceConnectorNexus(t *testing.T) {
 	})
 }
 
-func testAccDataSourceConnectorNexus(name string) string {
+func testAccDataSourceConnectorOciHelm(name string) string {
 	return fmt.Sprintf(`
-		resource "harness_platform_connector_nexus" "test" {
+	resource "harness_platform_secret_text" "test" {
+		identifier = "%[1]s"
+		name = "%[1]s"
+		description = "test"
+		tags = ["foo:bar"]
+
+		secret_manager_identifier = "harnessSecretManager"
+		value_type = "Inline"
+		value = "secret"
+	}
+
+		resource "harness_platform_connector_oci_helm" "test" {
 			identifier = "%[1]s"
 			name = "%[1]s"
 			description = "test"
 			tags = ["foo:bar"]
 
-			url = "https://nexus.example.com"
+			url = "admin.azurecr.io"
 			delegate_selectors = ["harness-delegate"]
-			version = "3.x"
 			credentials {
 				username = "admin"
-				password_ref = "account.doNotDeleteHSM"
+				password_ref = "account.${harness_platform_secret_text.test.id}"
 			}
+			depends_on = [time_sleep.wait_4_seconds]
 		}
 
-		data "harness_platform_connector_nexus" "test" {
-			identifier = harness_platform_connector_nexus.test.identifier
+		resource "time_sleep" "wait_4_seconds" {
+			depends_on = [harness_platform_secret_text.test]
+			destroy_duration = "4s"
+		}
+
+		data "harness_platform_connector_oci_helm" "test" {
+			identifier = harness_platform_connector_oci_helm.test.identifier
 		}
 	`, name)
 }
