+++ internal/service/cd_nextgen/connector/artifactRepositories/oci_helm_test.go
@@ -1,4 +1,4 @@
-package connector_test
+package artifactRepositories_test
 
 import (
 	"fmt"
